Lots more dynamokt updates including improvements to serialization.

+++ test/src/jvmMain/kotlin/com/steamstreet/aws/test/dynamo.kt
@@ -290,7 +290,7 @@ class PipesConfiguration(
             record.awsRegion,
             record.dynamodb.let {
                 DynamoStreamEventDetail(
-                    it.approximateCreationDateTime.time,
+                    it.approximateCreationDateTime.time.toDouble(),
                     srcRecord.dynamodb().keys(),
                     srcRecord.dynamodb().newImage(),
                     srcRecord.dynamodb().oldImage()
